test(react-grab): address review — extract harness, fix comments, settle waits
- Move FreezeHookHarness into its own module (mirrors PerfGrid), keeping
  App.tsx a composition root and relocating the external-store singleton.
- Correct the misleading "bypasses useState" comments (transition/context
  counters update via useState) and document the synthetic-click quirk.
- Add settle waits so the freeze-hold assertion can catch a leaked update
  and the post-unfreeze baseline is stable; guard readCount against a
  missing node; narrow the combined test to the simultaneity claim.
